Moved RFC exceptions into common library.
This makes them accessible to our authentication frameworks.
