+++ kangaroo-common/src/main/java/net/krotscheck/kangaroo/common/exception/rfc6749/package-info.java
@@ -18,4 +18,4 @@
 /**
  * Exception thrown by the OAuth API.
  */
-package net.krotscheck.api.oauth.exception.exception;
+package net.krotscheck.kangaroo.common.exception.rfc6749;
