+++ kangaroo-common/src/test/java/net/krotscheck/kangaroo/common/exception/rfc6749/ErrorCodeTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package net.krotscheck.api.oauth.exception.exception;
+package net.krotscheck.kangaroo.common.exception.rfc6749;
 
 import org.junit.Assert;
 import org.junit.Test;
