+++ kangaroo-server/src/main/java/net/krotscheck/api/oauth/resource/TokenService.java
@@ -18,7 +18,7 @@
 package net.krotscheck.api.oauth.resource;
 
 import net.krotscheck.api.oauth.annotation.OAuthFilterChain;
-import net.krotscheck.api.oauth.exception.exception.Rfc6749Exception.InvalidGrantException;
+import net.krotscheck.kangaroo.common.exception.rfc6749.Rfc6749Exception.InvalidGrantException;
 import net.krotscheck.api.oauth.factory.CredentialsFactory.Credentials;
 import net.krotscheck.api.oauth.resource.grant.IGrantTypeHandler;
 import net.krotscheck.kangaroo.database.entity.Client;
