+++ kangaroo-server/src/main/java/net/krotscheck/api/oauth/resource/grant/AuthorizationCodeGrantHandler.java
@@ -17,8 +17,8 @@
 
 package net.krotscheck.api.oauth.resource.grant;
 
-import net.krotscheck.api.oauth.exception.exception.Rfc6749Exception.InvalidGrantException;
-import net.krotscheck.api.oauth.exception.exception.Rfc6749Exception.InvalidRequestException;
+import net.krotscheck.kangaroo.common.exception.rfc6749.Rfc6749Exception.InvalidGrantException;
+import net.krotscheck.kangaroo.common.exception.rfc6749.Rfc6749Exception.InvalidRequestException;
 import net.krotscheck.api.oauth.resource.TokenResponseEntity;
 import net.krotscheck.kangaroo.database.entity.Client;
 import net.krotscheck.kangaroo.database.entity.ClientType;
