+++ kangaroo-server/src/test/java/net/krotscheck/api/oauth/resource/grant/ClientCredentialsGrantHandlerTest.java
@@ -17,9 +17,9 @@
 
 package net.krotscheck.api.oauth.resource.grant;
 
-import net.krotscheck.api.oauth.exception.exception.Rfc6749Exception.InvalidGrantException;
-import net.krotscheck.api.oauth.exception.exception.Rfc6749Exception.InvalidScopeException;
-import net.krotscheck.api.oauth.exception.exception.Rfc6749Exception.UnauthorizedClientException;
+import net.krotscheck.kangaroo.common.exception.rfc6749.Rfc6749Exception.InvalidGrantException;
+import net.krotscheck.kangaroo.common.exception.rfc6749.Rfc6749Exception.InvalidScopeException;
+import net.krotscheck.kangaroo.common.exception.rfc6749.Rfc6749Exception.UnauthorizedClientException;
 import net.krotscheck.api.oauth.resource.TokenResponseEntity;
 import net.krotscheck.kangaroo.database.entity.Client;
 import net.krotscheck.kangaroo.database.entity.ClientConfig;
